Define additional elements on other file (Fixes impossibility to edit form in VS2022)
