+++ NowNotes Windows/AdditionalElements.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Drawing.Drawing2D;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace NowNotes_Windows
+{
+	public class RoundButton : System.Windows.Forms.Button
+	{
+		protected override void OnPaint(System.Windows.Forms.PaintEventArgs e)
+		{
+			GraphicsPath grPath = new GraphicsPath();
+			grPath.AddEllipse(0, 0, ClientSize.Width, ClientSize.Height);
+			this.Region = new System.Drawing.Region(grPath);
+			base.OnPaint(e);
+		}
+	}
+}
